+++ internal/services/mobilenetwork/mobile_network_site_resource_test.go
@@ -22,9 +22,6 @@ type MobileNetworkSiteResource struct{}
 func TestAccMobileNetworkSite_basic(t *testing.T) {
 	data := acceptance.BuildTestData(t, "azurerm_mobile_network_site", "test")
 
-	// Limited regional availability for Mobile Network
-	data.Locations.Primary = "eastus"
-
 	r := MobileNetworkSiteResource{}
 	data.ResourceTest(t, r, []acceptance.TestStep{
 		{
@@ -39,9 +36,6 @@ func TestAccMobileNetworkSite_basic(t *testing.T) {
 
 func TestAccMobileNetworkSite_requiresImport(t *testing.T) {
 	data := acceptance.BuildTestData(t, "azurerm_mobile_network_site", "test")
-
-	// Limited regional availability for Mobile Network
-	data.Locations.Primary = "eastus"
 
 	r := MobileNetworkSiteResource{}
 	data.ResourceTest(t, r, []acceptance.TestStep{
@@ -58,9 +52,6 @@ func TestAccMobileNetworkSite_requiresImport(t *testing.T) {
 func TestAccMobileNetworkSite_complete(t *testing.T) {
 	data := acceptance.BuildTestData(t, "azurerm_mobile_network_site", "test")
 
-	// Limited regional availability for Mobile Network
-	data.Locations.Primary = "eastus"
-
 	r := MobileNetworkSiteResource{}
 	data.ResourceTest(t, r, []acceptance.TestStep{
 		{
@@ -75,9 +66,6 @@ func TestAccMobileNetworkSite_complete(t *testing.T) {
 
 func TestAccMobileNetworkSite_update(t *testing.T) {
 	data := acceptance.BuildTestData(t, "azurerm_mobile_network_site", "test")
-
-	// Limited regional availability for Mobile Network
-	data.Locations.Primary = "eastus"
 
 	r := MobileNetworkSiteResource{}
 	data.ResourceTest(t, r, []acceptance.TestStep{