+++ internal/services/network/client/client.go
@@ -6,17 +6,18 @@ package client
 import (
 	"fmt"
 
-	network_2023_02_01 "github.com/hashicorp/go-azure-sdk/resource-manager/network/2023-02-01"
+	network_2023_04_01 "github.com/hashicorp/go-azure-sdk/resource-manager/network/2023-04-01"
 	"github.com/hashicorp/go-azure-sdk/sdk/client/resourcemanager"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/common"
 	"github.com/tombuildsstuff/kermit/sdk/network/2022-07-01/network"
 )
 
 type Client struct {
-	*network_2023_02_01.Client
+	*network_2023_04_01.Client
 
 	ApplicationGatewaysClient              *network.ApplicationGatewaysClient
 	ConfigurationPolicyGroupClient         *network.ConfigurationPolicyGroupsClient
+	CustomIPPrefixesClient                 *network.CustomIPPrefixesClient
 	DDOSProtectionPlansClient              *network.DdosProtectionPlansClient
 	ExpressRouteAuthsClient                *network.ExpressRouteCircuitAuthorizationsClient
 	ExpressRouteCircuitsClient             *network.ExpressRouteCircuitsClient
@@ -71,6 +72,9 @@ func NewClient(o *common.ClientOptions) (*Client, error) {
 
 	configurationPolicyGroupClient := network.NewConfigurationPolicyGroupsClientWithBaseURI(o.ResourceManagerEndpoint, o.SubscriptionId)
 	o.ConfigureClient(&configurationPolicyGroupClient.Client, o.ResourceManagerAuthorizer)
+
+	customIpPrefixesClient := network.NewCustomIPPrefixesClient(o.SubscriptionId)
+	o.ConfigureClient(&customIpPrefixesClient.Client, o.ResourceManagerAuthorizer)
 
 	DDOSProtectionPlansClient := network.NewDdosProtectionPlansClientWithBaseURI(o.ResourceManagerEndpoint, o.SubscriptionId)
 	o.ConfigureClient(&DDOSProtectionPlansClient.Client, o.ResourceManagerAuthorizer)
@@ -210,7 +214,7 @@ func NewClient(o *common.ClientOptions) (*Client, error) {
 	ResourceNavigationLinkClient := network.NewResourceNavigationLinksClientWithBaseURI(o.ResourceManagerEndpoint, o.SubscriptionId)
 	o.ConfigureClient(&ResourceNavigationLinkClient.Client, o.ResourceManagerAuthorizer)
 
-	client, err := network_2023_02_01.NewClientWithBaseURI(o.Environment.ResourceManager, func(c *resourcemanager.Client) {
+	client, err := network_2023_04_01.NewClientWithBaseURI(o.Environment.ResourceManager, func(c *resourcemanager.Client) {
 		o.Configure(c, o.Authorizers.ResourceManager)
 	})
 	if err != nil {
@@ -222,6 +226,7 @@ func NewClient(o *common.ClientOptions) (*Client, error) {
 
 		ApplicationGatewaysClient:              &ApplicationGatewaysClient,
 		ConfigurationPolicyGroupClient:         &configurationPolicyGroupClient,
+		CustomIPPrefixesClient:                 &customIpPrefixesClient,
 		DDOSProtectionPlansClient:              &DDOSProtectionPlansClient,
 		ExpressRouteAuthsClient:                &ExpressRouteAuthsClient,
 		ExpressRouteCircuitsClient:             &ExpressRouteCircuitsClient,